+++ src/wirestate/core/error/error-code.ts
@@ -8,6 +8,7 @@ export const ERROR_CODE_NOT_ABSTRACT_SERVICE: number = 54;
 
 export const ERROR_CODE_FAILED_TO_RESOLVE: number = 100;
 export const ERROR_CODE_FAILED_TO_RESOLVE_QUERY_HANDLER: number = 101;
+export const ERROR_CODE_FAILED_TO_RESOLVE_COMMAND_HANDLER: number = 102;
 
 export const ERROR_CODE_ACCESS_BEFORE_ACTIVATION: number = 200;
 export const ERROR_CODE_ACCESS_AFTER_DISPOSAL: number = 201;