+++ src/wirestate/index.ts
@@ -7,6 +7,8 @@ export { bindEntry } from "@/wirestate/core/bind/bind-entry";
 export { bindConstant } from "@/wirestate/core/bind/bind-constant";
 export { bindService } from "@/wirestate/core/bind/bind-service";
 export { createIocContainer } from "@/wirestate/core/container/create-ioc-container";
+export { command } from "@/wirestate/core/container/command";
+export { commandOptional } from "@/wirestate/core/container/command-optional";
 export { emitSignal } from "@/wirestate/core/container/emit-signal";
 export { query } from "@/wirestate/core/container/query";
 export { queryOptional } from "@/wirestate/core/container/query-optional";
@@ -19,6 +21,10 @@ export {
 export { IocProvider } from "@/wirestate/core/provision/ioc-provider";
 export { useContainer } from "@/wirestate/core/provision/use-container";
 export { useContainerRevision } from "@/wirestate/core/provision/use-container-revision";
+export { OnCommand } from "@/wirestate/core/commands/on-command";
+export { useCommandCaller } from "@/wirestate/core/commands/use-command-caller";
+export { useOptionalCommandCaller } from "@/wirestate/core/commands/use-optional-command-caller";
+export { useCommandHandler } from "@/wirestate/core/commands/use-command-handler";
 export { OnQuery } from "@/wirestate/core/queries/on-query";
 export { useQueryCaller } from "@/wirestate/core/queries/use-query-caller";
 export { useOptionalQueryCaller } from "@/wirestate/core/queries/use-optional-query-caller";
@@ -42,6 +48,14 @@ export {
   TSeedKey as SeedKey,
 } from "@/wirestate/types/initial-state";
 export { IInjectableDescriptor as InjectableDescriptor } from "@/wirestate/types/privision";
+export {
+  ECommandStatus as CommandStatus,
+  ICommandDescriptor as CommandDescriptor,
+  TCommandHandler as CommandHandler,
+  TCommandType as CommandType,
+  TCommandUnregister as CommandUnregister,
+  TCommandCaller as CommandCaller,
+} from "@/wirestate/types/commands";
 export {
   TQueryHandler as QueryHandler,
   TQueryResponder as QueryResponder,