+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - `type` -> `bindingType`
 - `scopeType` -> `bindingScopeType`
 - Emit signal with two params instead of manually composed object, signal `from`
+- Add commands module
 
 ## 0.3.0